added input option + refactor
